+++ app/Modules/MailQueue/Support/MailQueue.php
@@ -23,12 +23,13 @@ class MailQueue
         return $object->mailQueue()->create([
             'from'       => json_encode(['email' => $object->user->email, 'name' => $object->user->name]),
             'to'         => json_encode($input['to']),
-            'cc'         => json_encode(($input['cc']) ?: []),
-            'bcc'        => json_encode(($input['bcc']) ?: []),
+            'cc'         => (isset($input['cc'])) ? json_encode($input['cc']) : json_encode(['']),
+            'bcc'        => (isset($input['bcc'])) ? json_encode($input['bcc']) :json_encode(['']),
             'subject'    => $input['subject'],
             'body'       => $input['body'],
             'attach_pdf' => $input['attach_pdf'],
         ]);
+
     }
 
     public function send($id)