+++ app/Widgets/Dashboard/QuoteSummary/Views/QuoteSummaryWidget.blade.php
@@ -16,14 +16,8 @@
 
             });
 
-            $('#quote-dashboard-total-setting-from-date').datepicker({
-                format: 'yyyy-mm-dd',
-                autoclose: true
-            });
-            $('#quote-dashboard-total-setting-to-date').datepicker({
-                format: 'yyyy-mm-dd',
-                autoclose: true
-            });
+            $('#quote-dashboard-total-setting-from-date').datepicker({format: 'yyyy-mm-dd', autoclose: true});
+            $('#quote-dashboard-total-setting-to-date').datepicker({format: 'yyyy-mm-dd', autoclose: true});
         });
     </script>
     <div class="card">